requirements.txt
arrow==1.2.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
dateparser==1.1.1
Flask==2.2.0
Flask-Cors==3.0.10
geographiclib==1.52
geojson==2.5.0
geopy==2.2.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-speech==2.15.0
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
ics==0.7.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.21
pyjokes==0.6.0
pyobjc==8.5
pyobjc-core==8.5
pyobjc-framework-Accessibility==8.5
pyobjc-framework-Accounts==8.5
pyobjc-framework-AddressBook==8.5
pyobjc-framework-AdServices==8.5
pyobjc-framework-AdSupport==8.5
pyobjc-framework-AppleScriptKit==8.5
pyobjc-framework-AppleScriptObjC==8.5
pyobjc-framework-ApplicationServices==8.5
pyobjc-framework-AppTrackingTransparency==8.5
pyobjc-framework-AudioVideoBridging==8.5
pyobjc-framework-AuthenticationServices==8.5
pyobjc-framework-AutomaticAssessmentConfiguration==8.5
pyobjc-framework-Automator==8.5
pyobjc-framework-AVFoundation==8.5
pyobjc-framework-AVKit==8.5
pyobjc-framework-BusinessChat==8.5
pyobjc-framework-CalendarStore==8.5
pyobjc-framework-CallKit==8.5
pyobjc-framework-CFNetwork==8.5
pyobjc-framework-ClassKit==8.5
pyobjc-framework-CloudKit==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Collaboration==8.5
pyobjc-framework-ColorSync==8.5
pyobjc-framework-Contacts==8.5
pyobjc-framework-ContactsUI==8.5
pyobjc-framework-CoreAudio==8.5
pyobjc-framework-CoreAudioKit==8.5
pyobjc-framework-CoreBluetooth==8.5
pyobjc-framework-CoreData==8.5
pyobjc-framework-CoreHaptics==8.5
pyobjc-framework-CoreLocation==8.5
pyobjc-framework-CoreMedia==8.5
pyobjc-framework-CoreMediaIO==8.5
pyobjc-framework-CoreMIDI==8.5
pyobjc-framework-CoreML==8.5
pyobjc-framework-CoreMotion==8.5
pyobjc-framework-CoreServices==8.5
pyobjc-framework-CoreSpotlight==8.5
pyobjc-framework-CoreText==8.5
pyobjc-framework-CoreWLAN==8.5
pyobjc-framework-CryptoTokenKit==8.5
pyobjc-framework-DataDetection==8.5
pyobjc-framework-DeviceCheck==8.5
pyobjc-framework-DictionaryServices==8.5
pyobjc-framework-DiscRecording==8.5
pyobjc-framework-DiscRecordingUI==8.5
pyobjc-framework-DiskArbitration==8.5
pyobjc-framework-DVDPlayback==8.5
pyobjc-framework-EventKit==8.5
pyobjc-framework-ExceptionHandling==8.5
pyobjc-framework-ExecutionPolicy==8.5
pyobjc-framework-ExternalAccessory==8.5
pyobjc-framework-FileProvider==8.5
pyobjc-framework-FileProviderUI==8.5
pyobjc-framework-FinderSync==8.5
pyobjc-framework-FSEvents==8.5
pyobjc-framework-GameCenter==8.5
pyobjc-framework-GameController==8.5
pyobjc-framework-GameKit==8.5
pyobjc-framework-GameplayKit==8.5
pyobjc-framework-ImageCaptureCore==8.5
pyobjc-framework-IMServicePlugIn==8.5
pyobjc-framework-InputMethodKit==8.5
pyobjc-framework-InstallerPlugins==8.5
pyobjc-framework-InstantMessage==8.5
pyobjc-framework-Intents==8.5
pyobjc-framework-IntentsUI==8.5
pyobjc-framework-IOSurface==8.5
pyobjc-framework-iTunesLibrary==8.5
pyobjc-framework-KernelManagement==8.5
pyobjc-framework-LatentSemanticMapping==8.5
pyobjc-framework-LaunchServices==8.5
pyobjc-framework-libdispatch==8.5
pyobjc-framework-LinkPresentation==8.5
pyobjc-framework-LocalAuthentication==8.5
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.5
pyobjc-framework-MailKit==8.5
pyobjc-framework-MapKit==8.5
pyobjc-framework-MediaAccessibility==8.5
pyobjc-framework-MediaLibrary==8.5
pyobjc-framework-MediaPlayer==8.5
pyobjc-framework-MediaToolbox==8.5
pyobjc-framework-Metal==8.5
pyobjc-framework-MetalKit==8.5
pyobjc-framework-MetalPerformanceShaders==8.5
pyobjc-framework-MetalPerformanceShadersGraph==8.5
pyobjc-framework-MetricKit==8.5
pyobjc-framework-MLCompute==8.5
pyobjc-framework-ModelIO==8.5
pyobjc-framework-MultipeerConnectivity==8.5
pyobjc-framework-NaturalLanguage==8.5
pyobjc-framework-NetFS==8.5
pyobjc-framework-Network==8.5
pyobjc-framework-NetworkExtension==8.5
pyobjc-framework-NotificationCenter==8.5
pyobjc-framework-OpenDirectory==8.5
pyobjc-framework-OSAKit==8.5
pyobjc-framework-OSLog==8.5
pyobjc-framework-PassKit==8.5
pyobjc-framework-PencilKit==8.5
pyobjc-framework-Photos==8.5
pyobjc-framework-PhotosUI==8.5
pyobjc-framework-PreferencePanes==8.5
pyobjc-framework-PushKit==8.5
pyobjc-framework-Quartz==8.5
pyobjc-framework-QuickLookThumbnailing==8.5
pyobjc-framework-ReplayKit==8.5
pyobjc-framework-SafariServices==8.5
pyobjc-framework-SceneKit==8.5
pyobjc-framework-ScreenCaptureKit==8.5
pyobjc-framework-ScreenSaver==8.5
pyobjc-framework-ScreenTime==8.5
pyobjc-framework-ScriptingBridge==8.5
pyobjc-framework-SearchKit==8.5
pyobjc-framework-Security==8.5
pyobjc-framework-SecurityFoundation==8.5
pyobjc-framework-SecurityInterface==8.5
pyobjc-framework-ServiceManagement==8.5
pyobjc-framework-ShazamKit==8.5
pyobjc-framework-Social==8.5
pyobjc-framework-SoundAnalysis==8.5
pyobjc-framework-Speech==8.5
pyobjc-framework-SpriteKit==8.5
pyobjc-framework-StoreKit==8.5
pyobjc-framework-SyncServices==8.5
pyobjc-framework-SystemConfiguration==8.5
pyobjc-framework-SystemExtensions==8.5
pyobjc-framework-UniformTypeIdentifiers==8.5
pyobjc-framework-UserNotifications==8.5
pyobjc-framework-UserNotificationsUI==8.5
pyobjc-framework-VideoSubscriberAccount==8.5
pyobjc-framework-VideoToolbox==8.5
pyobjc-framework-Virtualization==8.5
pyobjc-framework-Vision==8.5
pyobjc-framework-WebKit==8.5
pyowm==3.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
randfacts==0.18.0
regex==2022.3.2
requests==2.28.1
rsa==4.9
six==1.16.0
SpeechRecognition==3.8.1
srt==3.5.2
TatSu==5.8.3
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
vosk==0.3.43
Werkzeug==2.2.1
